pyproject.toml
[project]
name = "cryptonestx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-jazzmin>=3.0.1",
    "django-recaptcha>=4.1.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "six>=1.17.0",
]
